pyproject.toml
[project]
name = "recomusic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "click>=8.3.2",
    "fastapi>=0.136.0",
    "librosa>=0.11.0",
    "mutagen>=1.47.0",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.44.0",
]
